Refactor getting the latest dataset code
