+++ iyp/crawlers/stanford/asdb.py
@@ -1,14 +1,28 @@
+import re
 import sys
 import logging
 import requests
 import csv
+import bs4
+from bs4 import BeautifulSoup
+from datetime import datetime
 from iyp import BaseCrawler
-from get_latest_asdb_dataset import get_latest_asdb_dataset_url
+import bs4
+from bs4 import BeautifulSoup
 
-# TODO automate the file date
-INITIAL_URL = 'https://asdb.stanford.edu/#data'
-FILE_PATTERN = '%Y-%m_categorized_ases.csv'
-URL = get_latest_asdb_dataset_url(INITIAL_URL, FILE_PATTERN)
+def get_latest_asdb_dataset_url(asdb_stanford_data_url: str, file_name_format: str):
+    response = requests.get(asdb_stanford_data_url)
+    soup = BeautifulSoup(response.text, 'html.parser')
+    latest_date_element: bs4.element.Tag = soup.find("div",class_="col-md-12").find("p")
+    date_regex = re.compile(r'\d{1,2}/\d{1,2}/\d{4}')
+    date_string: str = date_regex.search(latest_date_element.text).group()
+    date: datetime = datetime.strptime(date_string, '%m/%d/%Y')
+    dateset_file_name: str = date.strftime(file_name_format)
+    asdb_stanford_data_url: str = asdb_stanford_data_url.replace('#', '')
+    full_url: str = f'{asdb_stanford_data_url}/{dateset_file_name}'
+    return full_url
+
+URL = get_latest_asdb_dataset_url('https://asdb.stanford.edu/#data', '%Y-%m_categorized_ases.csv')
 ORG = 'Stanford'
 NAME = 'stanford.asdb'
